Implemented processing seeds

+++ src/Interfaces/ISeedProducing.cs
@@ -3,6 +3,6 @@ namespace Trestlebridge.Interfaces
     public interface ISeedProducing
     {
         string Type { get; }
-        double Harvest();
+        int Harvest();
     }
 }
